+++ eclair-core/src/test/scala/fr/acinq/eclair/integration/ChannelIntegrationSpec.scala
@@ -1,0 +1,659 @@
+/*
+ * Copyright 2019 ACINQ SAS
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package fr.acinq.eclair.integration
+
+import java.util.UUID
+
+import akka.actor.ActorRef
+import akka.pattern.pipe
+import akka.testkit.TestProbe
+import com.typesafe.config.ConfigFactory
+import fr.acinq.bitcoin.Crypto.PublicKey
+import fr.acinq.bitcoin.{Base58, Base58Check, Bech32, ByteVector32, Crypto, OP_0, OP_CHECKSIG, OP_DUP, OP_EQUAL, OP_EQUALVERIFY, OP_HASH160, OP_PUSHDATA, Script, ScriptFlags, Transaction}
+import fr.acinq.eclair.blockchain.bitcoind.BitcoindService.BitcoinReq
+import fr.acinq.eclair.blockchain.bitcoind.rpc.ExtendedBitcoinClient
+import fr.acinq.eclair.channel._
+import fr.acinq.eclair.crypto.Sphinx.DecryptedFailurePacket
+import fr.acinq.eclair.io.Peer
+import fr.acinq.eclair.payment._
+import fr.acinq.eclair.payment.receive.MultiPartHandler.ReceivePayment
+import fr.acinq.eclair.payment.receive.{ForwardHandler, PaymentHandler}
+import fr.acinq.eclair.payment.send.PaymentInitiator.SendPaymentRequest
+import fr.acinq.eclair.transactions.{Scripts, Transactions}
+import fr.acinq.eclair.wire.{ChannelAnnouncement, PermanentChannelFailure, UpdateAddHtlc}
+import fr.acinq.eclair.{LongToBtcAmount, MilliSatoshi, randomBytes32}
+import org.json4s.JsonAST.{JString, JValue}
+import scodec.bits.ByteVector
+
+import scala.concurrent.ExecutionContext.Implicits.global
+import scala.concurrent.duration._
+import scala.jdk.CollectionConverters._
+
+/**
+ * Created by t-bast on 21/09/2020.
+ */
+
+class ChannelIntegrationSpec extends IntegrationSpec {
+
+  test("start eclair nodes") {
+    instantiateEclairNode("A", ConfigFactory.parseMap(Map("eclair.node-alias" -> "A", "eclair.expiry-delta-blocks" -> 130, "eclair.server.port" -> 29740, "eclair.api.port" -> 28090).asJava).withFallback(commonFeatures).withFallback(commonConfig))
+    instantiateEclairNode("C", ConfigFactory.parseMap(Map("eclair.node-alias" -> "C", "eclair.expiry-delta-blocks" -> 132, "eclair.server.port" -> 29741, "eclair.api.port" -> 28091).asJava).withFallback(withAnchorOutputs).withFallback(withWumbo).withFallback(commonConfig))
+    instantiateEclairNode("F1", ConfigFactory.parseMap(Map("eclair.node-alias" -> "F1", "eclair.expiry-delta-blocks" -> 135, "eclair.server.port" -> 29742, "eclair.api.port" -> 28092).asJava).withFallback(withWumbo).withFallback(commonConfig))
+    instantiateEclairNode("F2", ConfigFactory.parseMap(Map("eclair.node-alias" -> "F2", "eclair.expiry-delta-blocks" -> 136, "eclair.server.port" -> 29743, "eclair.api.port" -> 28093).asJava).withFallback(withAnchorOutputs).withFallback(commonConfig))
+  }
+
+  test("connect nodes") {
+    // A --- C --- {F1,F2}
+
+    val eventListener = TestProbe()
+    nodes("A").system.eventStream.subscribe(eventListener.ref, classOf[ChannelStateChanged])
+    nodes("C").system.eventStream.subscribe(eventListener.ref, classOf[ChannelStateChanged])
+
+    connect(nodes("A"), nodes("C"), 11000000 sat, 0 msat)
+    // confirm the funding tx
+    generateBlocks(bitcoincli, 2)
+    within(60 seconds) {
+      var count = 0
+      while (count < 2) {
+        if (eventListener.expectMsgType[ChannelStateChanged](30 seconds).currentState == NORMAL) count = count + 1
+      }
+    }
+
+    // generate more blocks so that all funding txes are buried under at least 6 blocks
+    generateBlocks(bitcoincli, 4)
+    awaitAnnouncements(1)
+  }
+
+  def awaitAnnouncements(channels: Int): Unit = {
+    val sender = TestProbe()
+    awaitCond({
+      sender.send(nodes("A").router, Symbol("channels"))
+      sender.expectMsgType[Iterable[ChannelAnnouncement]].size == channels
+    }, max = 60 seconds, interval = 1 second)
+  }
+
+  test("open channel C <-> F1 with push_msat and close") {
+    connect(nodes("C"), nodes("F1"), 5000000 sat, 500000000 msat)
+    generateBlocks(bitcoincli, 6)
+    awaitAnnouncements(2)
+
+    val sender = TestProbe()
+    sender.send(nodes("F1").register, Symbol("channelsTo"))
+    val channels = sender.expectMsgType[Map[ByteVector32, PublicKey]]
+    assert(channels.size === 1)
+    assert(channels.head._2 === nodes("C").nodeParams.nodeId)
+    val channelId = channels.head._1
+
+    sender.send(nodes("C").register, Register.Forward(sender.ref, channelId, CMD_GETSTATEDATA))
+    val commitmentsC = sender.expectMsgType[RES_GETSTATEDATA[DATA_NORMAL]].data.commitments
+    val finalPubKeyScriptC = commitmentsC.localParams.defaultFinalScriptPubKey
+    val fundingOutpoint = commitmentsC.commitInput.outPoint
+    sender.send(nodes("F1").register, Register.Forward(sender.ref, channelId, CMD_GETSTATEDATA))
+    val finalPubKeyScriptF = sender.expectMsgType[RES_GETSTATEDATA[DATA_NORMAL]].data.commitments.localParams.defaultFinalScriptPubKey
+
+    sender.send(nodes("F1").register, Register.Forward(sender.ref, channelId, CMD_CLOSE(Some(finalPubKeyScriptF))))
+    sender.expectMsgType[RES_SUCCESS[CMD_CLOSE]]
+    // we then wait for C and F to negotiate the closing fee
+    awaitCond({
+      sender.send(nodes("C").register, Register.Forward(sender.ref, channelId, CMD_GETSTATE))
+      sender.expectMsgType[RES_GETSTATE[State]].state == CLOSING
+    }, max = 20 seconds, interval = 1 second)
+
+    generateBlocks(bitcoincli, 2)
+    awaitCond({
+      sender.send(nodes("C").register, Register.Forward(sender.ref, channelId, CMD_GETSTATE))
+      sender.expectMsgType[RES_GETSTATE[State]].state == CLOSED
+    }, max = 20 seconds, interval = 1 second)
+
+    val bitcoinClient = new ExtendedBitcoinClient(bitcoinrpcclient)
+    bitcoinClient.lookForSpendingTx(None, fundingOutpoint.txid, fundingOutpoint.index.toInt).pipeTo(sender.ref)
+    val closingTx = sender.expectMsgType[Transaction]
+    assert(closingTx.txOut.map(_.publicKeyScript).toSet === Set(finalPubKeyScriptC, finalPubKeyScriptF))
+
+    awaitAnnouncements(1)
+  }
+
+  test("open channel C <-> F2, send payments and close (option_anchor_outputs, option_static_remotekey)") {
+    connect(nodes("C"), nodes("F2"), 5000000 sat, 0 msat)
+    generateBlocks(bitcoincli, 6)
+    awaitAnnouncements(2)
+
+    // initially all the balance is on C side and F2 doesn't have an output
+    val sender = TestProbe()
+    sender.send(nodes("F2").register, Symbol("channelsTo"))
+    // retrieve the channelId of C <--> F2
+    val Some(channelId) = sender.expectMsgType[Map[ByteVector32, PublicKey]].find(_._2 == nodes("C").nodeParams.nodeId).map(_._1)
+
+    sender.send(nodes("F2").register, Register.Forward(sender.ref, channelId, CMD_GETSTATEDATA))
+    val initialStateDataF2 = sender.expectMsgType[RES_GETSTATEDATA[DATA_NORMAL]].data
+    val initialCommitmentIndex = initialStateDataF2.commitments.localCommit.index
+
+    // the 'to remote' address is a simple script spending to the remote payment basepoint with a 1-block CSV delay
+    val toRemoteAddress = Script.pay2wsh(Scripts.toRemoteDelayed(initialStateDataF2.commitments.remoteParams.paymentBasepoint))
+
+    // toRemote output of C as seen by F2
+    val Some(toRemoteOutC) = initialStateDataF2.commitments.localCommit.publishableTxs.commitTx.tx.txOut.find(_.publicKeyScript == Script.write(toRemoteAddress))
+
+    // let's make a payment to advance the commit index
+    val amountMsat = 4200000.msat
+    sender.send(nodes("F2").paymentHandler, ReceivePayment(Some(amountMsat), "1 coffee"))
+    val pr = sender.expectMsgType[PaymentRequest]
+
+    // then we make the actual payment
+    sender.send(nodes("C").paymentInitiator, SendPaymentRequest(amountMsat, pr.paymentHash, nodes("F2").nodeParams.nodeId, maxAttempts = 1, fallbackFinalExpiryDelta = finalCltvExpiryDelta))
+    val paymentId = sender.expectMsgType[UUID](5 seconds)
+    val ps = sender.expectMsgType[PaymentSent](5 seconds)
+    assert(ps.id == paymentId)
+
+    sender.send(nodes("F2").register, Register.Forward(sender.ref, channelId, CMD_GETSTATEDATA))
+    val stateDataF2 = sender.expectMsgType[RES_GETSTATEDATA[DATA_NORMAL]].data
+    val commitmentIndex = stateDataF2.commitments.localCommit.index
+    val commitTx = stateDataF2.commitments.localCommit.publishableTxs.commitTx.tx
+    val Some(toRemoteOutCNew) = commitTx.txOut.find(_.publicKeyScript == Script.write(toRemoteAddress))
+
+    // there is a new commitment index in the channel state
+    assert(commitmentIndex == initialCommitmentIndex + 1)
+
+    // script pubkeys of toRemote output remained the same across commitments
+    assert(toRemoteOutCNew.publicKeyScript == toRemoteOutC.publicKeyScript)
+    assert(toRemoteOutCNew.amount < toRemoteOutC.amount)
+
+    // now let's force close the channel and check the toRemote is what we had at the beginning
+    sender.send(nodes("F2").register, Register.Forward(sender.ref, channelId, CMD_FORCECLOSE))
+    sender.expectMsgType[RES_SUCCESS[CMD_FORCECLOSE.type]]
+    // we then wait for C to detect the unilateral close and go to CLOSING state
+    awaitCond({
+      nodes("C").register ! Register.Forward(sender.ref, channelId, CMD_GETSTATE)
+      sender.expectMsgType[RES_GETSTATE[State]].state == CLOSING
+    }, max = 20 seconds, interval = 1 second)
+
+    val bitcoinClient = new ExtendedBitcoinClient(bitcoinrpcclient)
+    awaitCond({
+      bitcoinClient.getTransaction(commitTx.txid).pipeTo(sender.ref)
+      val tx = sender.expectMsgType[Transaction](10 seconds)
+      // the unilateral close contains the static toRemote output
+      tx.txOut.exists(_.publicKeyScript == toRemoteOutC.publicKeyScript)
+    }, max = 20 seconds, interval = 1 second)
+
+    // bury the unilateral close in a block, C should claim its main output
+    generateBlocks(bitcoincli, 2)
+    awaitCond({
+      bitcoinClient.getMempool().pipeTo(sender.ref)
+      sender.expectMsgType[Seq[Transaction]].exists(_.txIn.head.outPoint.txid === commitTx.txid)
+    }, max = 20 seconds, interval = 1 second)
+
+    // get the claim-remote-output confirmed, then the channel can go to the CLOSED state
+    generateBlocks(bitcoincli, 2)
+    awaitCond({
+      nodes("C").register ! Register.Forward(sender.ref, channelId, CMD_GETSTATE)
+      sender.expectMsgType[RES_GETSTATE[State]].state == CLOSED
+    }, max = 20 seconds, interval = 1 second)
+    awaitAnnouncements(1)
+  }
+
+  /**
+   * We currently use p2pkh script Helpers.getFinalScriptPubKey
+   */
+  def scriptPubKeyToAddress(scriptPubKey: ByteVector): String = Script.parse(scriptPubKey) match {
+    case OP_DUP :: OP_HASH160 :: OP_PUSHDATA(pubKeyHash, _) :: OP_EQUALVERIFY :: OP_CHECKSIG :: Nil =>
+      Base58Check.encode(Base58.Prefix.PubkeyAddressTestnet, pubKeyHash)
+    case OP_HASH160 :: OP_PUSHDATA(scriptHash, _) :: OP_EQUAL :: Nil =>
+      Base58Check.encode(Base58.Prefix.ScriptAddressTestnet, scriptHash)
+    case OP_0 :: OP_PUSHDATA(pubKeyHash, _) :: Nil if pubKeyHash.length == 20 => Bech32.encodeWitnessAddress("bcrt", 0, pubKeyHash)
+    case OP_0 :: OP_PUSHDATA(scriptHash, _) :: Nil if scriptHash.length == 32 => Bech32.encodeWitnessAddress("bcrt", 0, scriptHash)
+    case _ => ???
+  }
+
+  def listReceivedByAddress(address: String, sender: TestProbe = TestProbe()): Seq[ByteVector32] = {
+    sender.send(bitcoincli, BitcoinReq("listreceivedbyaddress", 0))
+    val res = sender.expectMsgType[JValue](10 seconds)
+    res.filter(_ \ "address" == JString(address)).flatMap(_ \ "txids" \\ classOf[JString]).map(ByteVector32.fromValidHex)
+  }
+
+  /** Wait for the given transaction to be either in the mempool or confirmed. */
+  def waitForTxBroadcastOrConfirmed(txid: ByteVector32, bitcoinClient: ExtendedBitcoinClient, sender: TestProbe): Unit = {
+    awaitCond({
+      bitcoinClient.getMempool().pipeTo(sender.ref)
+      val inMempool = sender.expectMsgType[Seq[Transaction]].exists(_.txid == txid)
+      bitcoinClient.getTxConfirmations(txid).pipeTo(sender.ref)
+      val confirmed = sender.expectMsgType[Option[Int]].nonEmpty
+      inMempool || confirmed
+    }, max = 30 seconds, interval = 1 second)
+  }
+
+  /** Disconnect node C from a given F node. */
+  def disconnectCF(nodeF: String, channelId: ByteVector32, sender: TestProbe = TestProbe()): Unit = {
+    sender.send(nodes(nodeF).switchboard, Symbol("peers"))
+    val peers = sender.expectMsgType[Iterable[ActorRef]]
+    // F's only node is C
+    peers.head ! Peer.Disconnect(nodes("C").nodeParams.nodeId)
+    // we then wait for F to be in disconnected state
+    awaitCond({
+      sender.send(nodes(nodeF).register, Register.Forward(sender.ref, channelId, CMD_GETSTATE))
+      sender.expectMsgType[RES_GETSTATE[State]].state == OFFLINE
+    }, max = 20 seconds, interval = 1 second)
+  }
+
+  case class ForceCloseFixture(sender: TestProbe, paymentSender: TestProbe, stateListener: TestProbe, paymentId: UUID, htlc: UpdateAddHtlc, preimage: ByteVector32, minerAddress: String, finalAddressC: String, finalAddressF: String)
+
+  /** Prepare a C <-> F channel for a force-close test by adding an HTLC that will be hodl-ed at F. */
+  def prepareForceCloseCF(nodeF: String, commitmentFormat: Transactions.CommitmentFormat): ForceCloseFixture = {
+    val sender = TestProbe()
+    sender.send(bitcoincli, BitcoinReq("getnewaddress"))
+    val JString(minerAddress) = sender.expectMsgType[JValue]
+    // we create and announce a channel between C and F; we use push_msat to ensure both nodes have an output in the commit tx
+    connect(nodes("C"), nodes(nodeF), 5000000 sat, 500000000 msat)
+    generateBlocks(bitcoincli, 6, Some(minerAddress))
+    awaitAnnouncements(2)
+    // we subscribe to C's channel state transitions
+    val stateListener = TestProbe()
+    nodes("C").system.eventStream.subscribe(stateListener.ref, classOf[ChannelStateChanged])
+    // first we make sure we are in sync with current blockchain height
+    val currentBlockCount = getBlockCount
+    awaitCond(getBlockCount == currentBlockCount, max = 20 seconds, interval = 1 second)
+    // we use this to control when to fulfill htlcs
+    val htlcReceiver = TestProbe()
+    nodes(nodeF).paymentHandler ! new ForwardHandler(htlcReceiver.ref)
+    val preimage = randomBytes32
+    val paymentHash = Crypto.sha256(preimage)
+    // A sends a payment to F
+    val paymentReq = SendPaymentRequest(100000000 msat, paymentHash, nodes(nodeF).nodeParams.nodeId, maxAttempts = 1, fallbackFinalExpiryDelta = finalCltvExpiryDelta, routeParams = integrationTestRouteParams)
+    val paymentSender = TestProbe()
+    paymentSender.send(nodes("A").paymentInitiator, paymentReq)
+    val paymentId = paymentSender.expectMsgType[UUID](30 seconds)
+    // F gets the htlc
+    val htlc = htlcReceiver.expectMsgType[IncomingPacket.FinalPacket].add
+    // now that we have the channel id, we retrieve channels default final addresses
+    sender.send(nodes("C").register, Register.Forward(sender.ref, htlc.channelId, CMD_GETSTATEDATA))
+    val dataC = sender.expectMsgType[RES_GETSTATEDATA[DATA_NORMAL]].data
+    assert(dataC.commitments.commitmentFormat === commitmentFormat)
+    val finalAddressC = scriptPubKeyToAddress(dataC.commitments.localParams.defaultFinalScriptPubKey)
+    sender.send(nodes(nodeF).register, Register.Forward(sender.ref, htlc.channelId, CMD_GETSTATEDATA))
+    val dataF = sender.expectMsgType[RES_GETSTATEDATA[DATA_NORMAL]].data
+    assert(dataF.commitments.commitmentFormat === commitmentFormat)
+    val finalAddressF = scriptPubKeyToAddress(dataF.commitments.localParams.defaultFinalScriptPubKey)
+    ForceCloseFixture(sender, paymentSender, stateListener, paymentId, htlc, preimage, minerAddress, finalAddressC, finalAddressF)
+  }
+
+  def testDownstreamFulfillLocalCommit(nodeF: String, commitmentFormat: Transactions.CommitmentFormat): Unit = {
+    val forceCloseFixture = prepareForceCloseCF(nodeF, commitmentFormat)
+    import forceCloseFixture._
+
+    // we retrieve transactions already received so that we don't take them into account when evaluating the outcome of this test
+    val previouslyReceivedByC = listReceivedByAddress(finalAddressC, sender)
+    // we then kill the connection between C and F
+    disconnectCF(nodeF, htlc.channelId, sender)
+    // we then have C unilaterally close the channel (which will make F redeem the htlc onchain)
+    sender.send(nodes("C").register, Register.Forward(sender.ref, htlc.channelId, CMD_FORCECLOSE))
+    sender.expectMsgType[RES_SUCCESS[CMD_FORCECLOSE.type]]
+    // we then wait for F to detect the unilateral close and go to CLOSING state
+    awaitCond({
+      sender.send(nodes(nodeF).register, Register.Forward(sender.ref, htlc.channelId, CMD_GETSTATE))
+      sender.expectMsgType[RES_GETSTATE[State]].state == CLOSING
+    }, max = 20 seconds, interval = 1 second)
+    // we generate a few blocks to get the commit tx confirmed
+    generateBlocks(bitcoincli, 3, Some(minerAddress))
+    // we then fulfill the htlc, which will make F redeem it on-chain
+    sender.send(nodes(nodeF).register, Register.Forward(sender.ref, htlc.channelId, CMD_FULFILL_HTLC(htlc.id, preimage)))
+    // we don't need to generate blocks to confirm the htlc-success; C should extract the preimage as soon as it enters
+    // the mempool and fulfill the payment upstream.
+    paymentSender.expectMsgType[PaymentSent](30 seconds)
+    // we then generate enough blocks so that nodes get their main delayed output
+    generateBlocks(bitcoincli, 145, Some(minerAddress))
+    // F should have 2 recv transactions: the redeemed htlc and its main output
+    // C should have 1 recv transaction: its main output
+    awaitCond({
+      val receivedByC = listReceivedByAddress(finalAddressC, sender)
+      val receivedByF = listReceivedByAddress(finalAddressF)
+      receivedByF.size == 2 && (receivedByC diff previouslyReceivedByC).size == 1
+    }, max = 30 seconds, interval = 1 second)
+    // we generate blocks to make txs confirm
+    generateBlocks(bitcoincli, 2, Some(minerAddress))
+    // and we wait for the channel to close
+    awaitCond(stateListener.expectMsgType[ChannelStateChanged].currentState == CLOSED, max = 30 seconds)
+    awaitCond({
+      sender.send(nodes(nodeF).register, Register.Forward(sender.ref, htlc.channelId, CMD_GETSTATE))
+      sender.expectMsgType[RES_GETSTATE[State]].state == CLOSED
+    }, max = 20 seconds, interval = 1 second)
+    awaitAnnouncements(1)
+  }
+
+  test("propagate a fulfill upstream when a downstream htlc is redeemed on-chain (local commit)") {
+    testDownstreamFulfillLocalCommit("F1", Transactions.DefaultCommitmentFormat)
+  }
+
+  test("propagate a fulfill upstream when a downstream htlc is redeemed on-chain (local commit, anchor outputs)") {
+    testDownstreamFulfillLocalCommit("F2", Transactions.AnchorOutputsCommitmentFormat)
+  }
+
+  def testDownstreamFulfillRemoteCommit(nodeF: String, commitmentFormat: Transactions.CommitmentFormat): Unit = {
+    val forceCloseFixture = prepareForceCloseCF(nodeF, commitmentFormat)
+    import forceCloseFixture._
+
+    // we retrieve transactions already received so that we don't take them into account when evaluating the outcome of this test
+    val previouslyReceivedByC = listReceivedByAddress(finalAddressC, sender)
+    // we then kill the connection between C and F
+    disconnectCF(nodeF, htlc.channelId, sender)
+    // then we have F unilaterally close the channel
+    sender.send(nodes(nodeF).register, Register.Forward(sender.ref, htlc.channelId, CMD_FORCECLOSE))
+    sender.expectMsgType[RES_SUCCESS[CMD_FORCECLOSE.type]]
+    awaitCond(stateListener.expectMsgType[ChannelStateChanged].currentState == CLOSING, max = 30 seconds)
+    // we generate a few blocks to get the commit tx confirmed
+    generateBlocks(bitcoincli, 3, Some(minerAddress))
+    // we then fulfill the htlc (it won't be sent to C, and will be used to pull funds on-chain)
+    sender.send(nodes(nodeF).register, Register.Forward(sender.ref, htlc.channelId, CMD_FULFILL_HTLC(htlc.id, preimage)))
+    // we don't need to generate blocks to confirm the htlc-success; C should extract the preimage as soon as it enters
+    // the mempool and fulfill the payment upstream.
+    paymentSender.expectMsgType[PaymentSent](30 seconds)
+    // we then generate enough blocks so that F gets its htlc-success delayed output
+    generateBlocks(bitcoincli, 145, Some(minerAddress))
+    // F should have 2 recv transactions: the redeemed htlc and its main output
+    // C should have 1 recv transaction: its main output
+    awaitCond({
+      val receivedByC = listReceivedByAddress(finalAddressC, sender)
+      val receivedByF = listReceivedByAddress(finalAddressF, sender)
+      receivedByF.size == 2 && (receivedByC diff previouslyReceivedByC).size == 1
+    }, max = 30 seconds, interval = 1 second)
+    // we generate blocks to make txs confirm
+    generateBlocks(bitcoincli, 2, Some(minerAddress))
+    // and we wait for the channel to close
+    awaitCond(stateListener.expectMsgType[ChannelStateChanged].currentState == CLOSED, max = 30 seconds)
+    awaitCond({
+      sender.send(nodes(nodeF).register, Register.Forward(sender.ref, htlc.channelId, CMD_GETSTATE))
+      sender.expectMsgType[RES_GETSTATE[State]].state == CLOSED
+    }, max = 20 seconds, interval = 1 second)
+    awaitAnnouncements(1)
+  }
+
+  test("propagate a fulfill upstream when a downstream htlc is redeemed on-chain (remote commit)") {
+    testDownstreamFulfillRemoteCommit("F1", Transactions.DefaultCommitmentFormat)
+  }
+
+  test("propagate a fulfill upstream when a downstream htlc is redeemed on-chain (remote commit, anchor outputs)") {
+    testDownstreamFulfillRemoteCommit("F2", Transactions.AnchorOutputsCommitmentFormat)
+  }
+
+  def testDownstreamTimeoutLocalCommit(nodeF: String, commitmentFormat: Transactions.CommitmentFormat): Unit = {
+    val forceCloseFixture = prepareForceCloseCF(nodeF, commitmentFormat)
+    import forceCloseFixture._
+
+    // we retrieve transactions already received so that we don't take them into account when evaluating the outcome of this test
+    val previouslyReceivedByC = listReceivedByAddress(finalAddressC, sender)
+    // we then kill the connection between C and F; otherwise F would send an error message to C when it detects the htlc
+    // timeout. When that happens C would broadcast his commit tx, and if it gets to the mempool before F's commit tx we
+    // won't be testing the right scenario.
+    disconnectCF(nodeF, htlc.channelId, sender)
+    // we generate enough blocks to reach the htlc timeout
+    generateBlocks(bitcoincli, (htlc.cltvExpiry.toLong - getBlockCount).toInt, Some(minerAddress))
+    awaitCond(stateListener.expectMsgType[ChannelStateChanged].currentState == CLOSING, max = 30 seconds)
+    sender.send(nodes("C").register, Register.Forward(sender.ref, htlc.channelId, CMD_GETSTATEDATA))
+    val Some(localCommit) = sender.expectMsgType[RES_GETSTATEDATA[DATA_CLOSING]].data.localCommitPublished
+    // we wait until the commit tx has been broadcast
+    val bitcoinClient = new ExtendedBitcoinClient(bitcoinrpcclient)
+    waitForTxBroadcastOrConfirmed(localCommit.commitTx.txid, bitcoinClient, sender)
+    // we generate a few blocks to get the commit tx confirmed
+    generateBlocks(bitcoincli, 3, Some(minerAddress))
+    // we wait until the htlc-timeout has been broadcast
+    waitForTxBroadcastOrConfirmed(localCommit.htlcTimeoutTxs.head.txid, bitcoinClient, sender)
+    // we generate more blocks for the htlc-timeout to reach enough confirmations
+    generateBlocks(bitcoincli, 3, Some(minerAddress))
+    // this will fail the htlc
+    val failed = paymentSender.expectMsgType[PaymentFailed](30 seconds)
+    assert(failed.id == paymentId)
+    assert(failed.paymentHash === htlc.paymentHash)
+    assert(failed.failures.nonEmpty)
+    assert(failed.failures.head.asInstanceOf[RemoteFailure].e === DecryptedFailurePacket(nodes("C").nodeParams.nodeId, PermanentChannelFailure))
+    // we then generate enough blocks to confirm all delayed transactions
+    generateBlocks(bitcoincli, 145, Some(minerAddress))
+    // C should have 2 recv transactions: its main output and the htlc timeout
+    // F should have 1 recv transaction: its main output
+    awaitCond({
+      val receivedByC = listReceivedByAddress(finalAddressC, sender)
+      val receivedByF = listReceivedByAddress(finalAddressF, sender)
+      receivedByF.size == 1 && (receivedByC diff previouslyReceivedByC).size == 2
+    }, max = 30 seconds, interval = 1 second)
+    // we generate blocks to make txs confirm
+    generateBlocks(bitcoincli, 2, Some(minerAddress))
+    // and we wait for the channel to close
+    awaitCond(stateListener.expectMsgType[ChannelStateChanged].currentState == CLOSED, max = 30 seconds)
+    awaitCond({
+      sender.send(nodes(nodeF).register, Register.Forward(sender.ref, htlc.channelId, CMD_GETSTATE))
+      sender.expectMsgType[RES_GETSTATE[State]].state == CLOSED
+    }, max = 20 seconds, interval = 1 second)
+    awaitAnnouncements(1)
+  }
+
+  test("propagate a failure upstream when a downstream htlc times out (local commit)") {
+    testDownstreamTimeoutLocalCommit("F1", Transactions.DefaultCommitmentFormat)
+  }
+
+  test("propagate a failure upstream when a downstream htlc times out (local commit, anchor outputs)") {
+    testDownstreamTimeoutLocalCommit("F2", Transactions.AnchorOutputsCommitmentFormat)
+  }
+
+  def testDownstreamTimeoutRemoteCommit(nodeF: String, commitmentFormat: Transactions.CommitmentFormat): Unit = {
+    val forceCloseFixture = prepareForceCloseCF(nodeF, commitmentFormat)
+    import forceCloseFixture._
+
+    // we retrieve transactions already received so that we don't take them into account when evaluating the outcome of this test
+    val previouslyReceivedByC = listReceivedByAddress(finalAddressC, sender)
+    // we then kill the connection between C and F to ensure the close can only be detected on-chain
+    disconnectCF(nodeF, htlc.channelId, sender)
+    // we ask F to unilaterally close the channel
+    sender.send(nodes(nodeF).register, Register.Forward(sender.ref, htlc.channelId, CMD_FORCECLOSE))
+    sender.expectMsgType[RES_SUCCESS[CMD_FORCECLOSE.type]]
+    // we wait for C to detect the unilateral close
+    awaitCond({
+      sender.send(nodes("C").register, Register.Forward(sender.ref, htlc.channelId, CMD_GETSTATEDATA))
+      sender.expectMsgType[RES_GETSTATEDATA[Data]].data match {
+        case d: DATA_CLOSING if d.remoteCommitPublished.nonEmpty => true
+        case _ => false
+      }
+    }, max = 30 seconds, interval = 1 second)
+    sender.send(nodes("C").register, Register.Forward(sender.ref, htlc.channelId, CMD_GETSTATEDATA))
+    val Some(remoteCommit) = sender.expectMsgType[RES_GETSTATEDATA[DATA_CLOSING]].data.remoteCommitPublished
+    // we generate enough blocks to make the htlc timeout
+    generateBlocks(bitcoincli, (htlc.cltvExpiry.toLong - getBlockCount).toInt, Some(minerAddress))
+    // we wait until the claim-htlc-timeout has been broadcast
+    val bitcoinClient = new ExtendedBitcoinClient(bitcoinrpcclient)
+    waitForTxBroadcastOrConfirmed(remoteCommit.claimHtlcTimeoutTxs.head.txid, bitcoinClient, sender)
+    // and we generate blocks for the claim-htlc-timeout to reach enough confirmations
+    generateBlocks(bitcoincli, 3, Some(minerAddress))
+    // this will fail the htlc
+    val failed = paymentSender.expectMsgType[PaymentFailed](30 seconds)
+    assert(failed.id == paymentId)
+    assert(failed.paymentHash === htlc.paymentHash)
+    assert(failed.failures.nonEmpty)
+    assert(failed.failures.head.asInstanceOf[RemoteFailure].e === DecryptedFailurePacket(nodes("C").nodeParams.nodeId, PermanentChannelFailure))
+    // we then generate enough blocks to confirm all delayed transactions
+    generateBlocks(bitcoincli, 145, Some(minerAddress))
+    // C should have 2 recv transactions: its main output and the htlc timeout
+    // F should have 1 recv transaction: its main output
+    awaitCond({
+      val receivedByC = listReceivedByAddress(finalAddressC, sender)
+      val receivedByF = listReceivedByAddress(finalAddressF, sender)
+      receivedByF.size == 1 && (receivedByC diff previouslyReceivedByC).size == 2
+    }, max = 30 seconds, interval = 1 second)
+    // we generate blocks to make tx confirm
+    generateBlocks(bitcoincli, 2, Some(minerAddress))
+    // and we wait for the channel to close
+    awaitCond(stateListener.expectMsgType[ChannelStateChanged].currentState == CLOSED, max = 30 seconds)
+    awaitCond({
+      sender.send(nodes(nodeF).register, Register.Forward(sender.ref, htlc.channelId, CMD_GETSTATE))
+      sender.expectMsgType[RES_GETSTATE[State]].state == CLOSED
+    }, max = 20 seconds, interval = 1 second)
+    awaitAnnouncements(1)
+  }
+
+  test("propagate a failure upstream when a downstream htlc times out (remote commit)") {
+    testDownstreamTimeoutRemoteCommit("F1", Transactions.DefaultCommitmentFormat)
+  }
+
+  test("propagate a failure upstream when a downstream htlc times out (remote commit, anchor outputs)") {
+    testDownstreamTimeoutRemoteCommit("F2", Transactions.AnchorOutputsCommitmentFormat)
+  }
+
+  case class RevokedCommitFixture(sender: TestProbe, stateListenerC: TestProbe, revokedCommitTx: Transaction, htlcSuccess: Seq[Transaction], htlcTimeout: Seq[Transaction], finalAddressC: String)
+
+  def testRevokedCommit(nodeF: String, commitmentFormat: Transactions.CommitmentFormat): RevokedCommitFixture = {
+    val sender = TestProbe()
+    // we create and announce a channel between C and F; we use push_msat to ensure F has a balance
+    connect(nodes("C"), nodes(nodeF), 5000000 sat, 300000000 msat)
+    generateBlocks(bitcoincli, 6)
+    awaitAnnouncements(2)
+    // we subscribe to C's channel state transitions
+    val stateListenerC = TestProbe()
+    nodes("C").system.eventStream.subscribe(stateListenerC.ref, classOf[ChannelStateChanged])
+    // we use this to get commitments
+    val sigListener = TestProbe()
+    nodes(nodeF).system.eventStream.subscribe(sigListener.ref, classOf[ChannelSignatureReceived])
+    // we use this to control when to fulfill htlcs
+    val forwardHandlerC = TestProbe()
+    nodes("C").paymentHandler ! new ForwardHandler(forwardHandlerC.ref)
+    val forwardHandlerF = TestProbe()
+    nodes(nodeF).paymentHandler ! new ForwardHandler(forwardHandlerF.ref)
+    // this is the actual payment handler that we will forward requests to
+    val paymentHandlerC = nodes("C").system.actorOf(PaymentHandler.props(nodes("C").nodeParams, nodes("C").register))
+    val paymentHandlerF = nodes(nodeF).system.actorOf(PaymentHandler.props(nodes(nodeF).nodeParams, nodes(nodeF).register))
+    // first we make sure nodes are in sync with current blockchain height
+    val currentBlockCount = getBlockCount
+    awaitCond(getBlockCount == currentBlockCount, max = 20 seconds, interval = 1 second)
+
+    // we now send a few htlcs C->F and F->C in order to obtain a commitments with multiple htlcs
+    def send(amountMsat: MilliSatoshi, paymentHandler: ActorRef, paymentInitiator: ActorRef): UUID = {
+      sender.send(paymentHandler, ReceivePayment(Some(amountMsat), "1 coffee"))
+      val pr = sender.expectMsgType[PaymentRequest]
+      val sendReq = SendPaymentRequest(amountMsat, pr.paymentHash, pr.nodeId, maxAttempts = 1, fallbackFinalExpiryDelta = finalCltvExpiryDelta, routeParams = integrationTestRouteParams)
+      sender.send(paymentInitiator, sendReq)
+      sender.expectMsgType[UUID]
+    }
+
+    val buffer = TestProbe()
+    send(100000000 msat, paymentHandlerF, nodes("C").paymentInitiator)
+    forwardHandlerF.expectMsgType[IncomingPacket.FinalPacket]
+    forwardHandlerF.forward(buffer.ref)
+    sigListener.expectMsgType[ChannelSignatureReceived]
+    send(110000000 msat, paymentHandlerF, nodes("C").paymentInitiator)
+    forwardHandlerF.expectMsgType[IncomingPacket.FinalPacket]
+    forwardHandlerF.forward(buffer.ref)
+    sigListener.expectMsgType[ChannelSignatureReceived]
+    send(120000000 msat, paymentHandlerC, nodes(nodeF).paymentInitiator)
+    forwardHandlerC.expectMsgType[IncomingPacket.FinalPacket]
+    forwardHandlerC.forward(buffer.ref)
+    sigListener.expectMsgType[ChannelSignatureReceived]
+    send(130000000 msat, paymentHandlerC, nodes(nodeF).paymentInitiator)
+    forwardHandlerC.expectMsgType[IncomingPacket.FinalPacket]
+    forwardHandlerC.forward(buffer.ref)
+    val commitmentsF = sigListener.expectMsgType[ChannelSignatureReceived].commitments
+    sigListener.expectNoMsg(1 second)
+    assert(commitmentsF.commitmentFormat === commitmentFormat)
+    // in this commitment, both parties should have a main output, there are four pending htlcs and anchor outputs if applicable
+    val localCommitF = commitmentsF.localCommit.publishableTxs
+    commitmentFormat match {
+      case Transactions.DefaultCommitmentFormat => assert(localCommitF.commitTx.tx.txOut.size === 6)
+      case Transactions.AnchorOutputsCommitmentFormat => assert(localCommitF.commitTx.tx.txOut.size === 8)
+    }
+    val htlcTimeoutTxs = localCommitF.htlcTxsAndSigs.collect { case h@HtlcTxAndSigs(_: Transactions.HtlcTimeoutTx, _, _) => h }
+    val htlcSuccessTxs = localCommitF.htlcTxsAndSigs.collect { case h@HtlcTxAndSigs(_: Transactions.HtlcSuccessTx, _, _) => h }
+    assert(htlcTimeoutTxs.size === 2)
+    assert(htlcSuccessTxs.size === 2)
+    // we fulfill htlcs to get the preimages
+    buffer.expectMsgType[IncomingPacket.FinalPacket]
+    buffer.forward(paymentHandlerF)
+    sigListener.expectMsgType[ChannelSignatureReceived]
+    val preimage1 = sender.expectMsgType[PaymentSent].paymentPreimage
+    buffer.expectMsgType[IncomingPacket.FinalPacket]
+    buffer.forward(paymentHandlerF)
+    sigListener.expectMsgType[ChannelSignatureReceived]
+    val preimage2 = sender.expectMsgType[PaymentSent].paymentPreimage
+    buffer.expectMsgType[IncomingPacket.FinalPacket]
+    buffer.forward(paymentHandlerC)
+    sigListener.expectMsgType[ChannelSignatureReceived]
+    sender.expectMsgType[PaymentSent]
+    buffer.expectMsgType[IncomingPacket.FinalPacket]
+    buffer.forward(paymentHandlerC)
+    sigListener.expectMsgType[ChannelSignatureReceived]
+    sender.expectMsgType[PaymentSent]
+    // we then generate blocks to make htlcs timeout (nothing will happen in the channel because all of them have already been fulfilled)
+    generateBlocks(bitcoincli, 40)
+    // we retrieve C's default final address
+    sender.send(nodes("C").register, Register.Forward(sender.ref, commitmentsF.channelId, CMD_GETSTATEDATA))
+    val finalAddressC = scriptPubKeyToAddress(sender.expectMsgType[RES_GETSTATEDATA[DATA_NORMAL]].data.commitments.localParams.defaultFinalScriptPubKey)
+    // we prepare the revoked transactions F will publish
+    val revokedCommitTx = localCommitF.commitTx.tx
+    val htlcSuccess = htlcSuccessTxs.zip(Seq(preimage1, preimage2)).map { case (htlcTxAndSigs, preimage) => Transactions.addSigs(htlcTxAndSigs.txinfo.asInstanceOf[Transactions.HtlcSuccessTx], htlcTxAndSigs.localSig, htlcTxAndSigs.remoteSig, preimage, commitmentsF.commitmentFormat).tx }
+    val htlcTimeout = htlcTimeoutTxs.map(htlcTxAndSigs => Transactions.addSigs(htlcTxAndSigs.txinfo.asInstanceOf[Transactions.HtlcTimeoutTx], htlcTxAndSigs.localSig, htlcTxAndSigs.remoteSig, commitmentsF.commitmentFormat).tx)
+    htlcSuccess.foreach(tx => Transaction.correctlySpends(tx, Seq(revokedCommitTx), ScriptFlags.STANDARD_SCRIPT_VERIFY_FLAGS))
+    htlcTimeout.foreach(tx => Transaction.correctlySpends(tx, Seq(revokedCommitTx), ScriptFlags.STANDARD_SCRIPT_VERIFY_FLAGS))
+    RevokedCommitFixture(sender, stateListenerC, revokedCommitTx, htlcSuccess, htlcTimeout, finalAddressC)
+  }
+
+  test("punish a node that has published a revoked commit tx") {
+    val revokedCommitFixture = testRevokedCommit("F1", Transactions.DefaultCommitmentFormat)
+    import revokedCommitFixture._
+
+    val bitcoinClient = new ExtendedBitcoinClient(bitcoinrpcclient)
+    // we retrieve transactions already received so that we don't take them into account when evaluating the outcome of this test
+    val previouslyReceivedByC = listReceivedByAddress(finalAddressC, sender)
+    // F publishes the revoked commitment, one HTLC-success, one HTLC-timeout and leaves the other HTLC outputs unclaimed
+    bitcoinClient.publishTransaction(revokedCommitTx).pipeTo(sender.ref)
+    sender.expectMsg(revokedCommitTx.txid)
+    bitcoinClient.publishTransaction(htlcSuccess.head).pipeTo(sender.ref)
+    sender.expectMsg(htlcSuccess.head.txid)
+    bitcoinClient.publishTransaction(htlcTimeout.head).pipeTo(sender.ref)
+    sender.expectMsg(htlcTimeout.head.txid)
+    // at this point C should have 6 recv transactions: its previous main output, F's main output and all htlc outputs (taken as punishment)
+    awaitCond({
+      val receivedByC = listReceivedByAddress(finalAddressC, sender)
+      (receivedByC diff previouslyReceivedByC).size == 6
+    }, max = 30 seconds, interval = 1 second)
+    // we generate blocks to make txs confirm
+    generateBlocks(bitcoincli, 2)
+    // and we wait for C's channel to close
+    awaitCond(stateListenerC.expectMsgType[ChannelStateChanged].currentState == CLOSED, max = 30 seconds)
+    awaitAnnouncements(1)
+  }
+
+  test("punish a node that has published a revoked commit tx (anchor outputs)") {
+    val revokedCommitFixture = testRevokedCommit("F2", Transactions.AnchorOutputsCommitmentFormat)
+    import revokedCommitFixture._
+
+    val bitcoinClient = new ExtendedBitcoinClient(bitcoinrpcclient)
+    // we retrieve transactions already received so that we don't take them into account when evaluating the outcome of this test
+    val previouslyReceivedByC = listReceivedByAddress(finalAddressC, sender)
+    // F publishes the revoked commitment: it can't publish the HTLC txs because of the CSV 1
+    bitcoinClient.publishTransaction(revokedCommitTx).pipeTo(sender.ref)
+    sender.expectMsg(revokedCommitTx.txid)
+    // get the revoked commitment confirmed: now HTLC txs can be published
+    generateBlocks(bitcoincli, 2)
+    bitcoinClient.publishTransaction(htlcSuccess.head).pipeTo(sender.ref)
+    sender.expectMsg(htlcSuccess.head.txid)
+    bitcoinClient.publishTransaction(htlcTimeout.head).pipeTo(sender.ref)
+    sender.expectMsg(htlcTimeout.head.txid)
+    // at this point C should have 6 recv transactions: its previous main output, F's main output and all htlc outputs (taken as punishment)
+    awaitCond({
+      val receivedByC = listReceivedByAddress(finalAddressC, sender)
+      (receivedByC diff previouslyReceivedByC).size == 6
+    }, max = 30 seconds, interval = 1 second)
+    // we generate blocks to make txs confirm
+    generateBlocks(bitcoincli, 2)
+    // and we wait for C's channel to close
+    awaitCond(stateListenerC.expectMsgType[ChannelStateChanged].currentState == CLOSED, max = 30 seconds)
+    awaitAnnouncements(1)
+  }
+
+}
